chore: rename files, consolidate graphql tests
